requirements.txt
Flask==3.0.2
psutil==5.9.8
python-dotenv==1.0.1 